+++ src/main/java/ru/danon/spring/ToDo/dto/NotificationEvent.java
@@ -10,14 +10,14 @@ public class NotificationEvent {
     private String type;
     private String title;
     private String message;
-    private Integer userId;
+    private Long userId;
     private String userRole;
     private Timestamp createdAt;
     private Map<String, Object> metadata;
 
     public NotificationEvent() {}
 
-    public NotificationEvent(String id, String type, String title, String message, Integer userId, String userRole, Timestamp createdAt, Map<String, Object> metadata) {
+    public NotificationEvent(String id, String type, String title, String message, Long userId, String userRole, Timestamp createdAt, Map<String, Object> metadata) {
         this.id = id;
         this.type = type;
         this.title = title;
