+++ src/main/java/ru/danon/spring/ToDo/dto/TaskFileDTO.java
@@ -6,12 +6,12 @@ import java.time.LocalDateTime;
 
 @Data
 public class TaskFileDTO {
-    private Integer id;
+    private Long id;
     private String originalFileName;
     private Long fileSize;
     private String fileType;
     private LocalDateTime uploadedAt;
-    private Integer uploadedById;
+    private Long uploadedById;
     private String uploadedByName;
     private String downloadUrl;
 }
